created core file with all the functionality
